+++ CHANGELOG.md
@@ -18,10 +18,12 @@
 - Cleared password conflicts and issues
 
 ## [2025-07-14]
-- Made the interactive UI for "Prescriptions at a Glance" page
-- Made the interactive UI for "Medicine Reminders" page
+- Coded the interactive UI for "Prescriptions at a Glance" page
+- Coded the interactive UI for "Medicine Reminders" page
 
 ## [2025-07-15]
-- Made the 'Scan a Prescription' page
+- Coded the 'Scan a Prescription' page
 - Integrated camera feature in the 'Scan Prescription' page
 - Added screenshots to README.md
+- Coded the 'My Profile' page
+- Added Screenshots of 'My Profile' page on README.md